requirements-test.txt
# Core Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
pytest-xdist==3.5.0
pytest-mock==3.12.0

# Factory for Test Data
factory-boy==3.3.0
faker==20.1.0

# Additional
pytest-timeout==2.2.0
pytest-randomly==3.15.0
